Triple Input/Output on STDIN/STDOUT
The generate triples script now reads the list of files from standard
in.  This is helpful to use the linux parallel command to use multiple
processors to do the computation in parallel.  Similarly, output now
goes to standard output so that it can be redirected to a file, which
allows parallel to correctly put the output together.  Since these are
triples, the order doesn't matter, as long as entire lines are written
at a time.